tu: support KHR_8bit_storage
Provide a separate descriptor for 8-bit storage access. The descriptor
index is adjusted appropriately for 8-bit SSBO loads and stores.

The 8-bit SSBO loads cannot go through isam since that instruction isn't
able to handle those. The ldib and stib instruction encodings are a bit
peculiar but they match the blob's image buffer access through VK_FORMAT_R8
and the dedicated descriptor. These loads and stores do not work in
vectorized form, so they have to be scalarized.

8-bit zero-extension and truncation do not function correctly when applied
through the cov instruction. Instead, the correct result is achieved by
masking. 8-bit sign extension through cov from a non-shared to a shared
register also doesn't work, so an exception is applied to avoid it.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/9979
Signed-off-by: Zan Dobersek <zdobersek@igalia.com>