aco/live_var_analysis: Fix fixed VGPRs intersecting linear VGPR bounds
If there is a fixed VGPR in a high register with low register demand
(for example a definition fixed to v31 with demand = 32), then a linear
VGPR might already be in that register. We can't move/split live ranges
of linear VGPRs, so take the maximum fixed VGPR into account when
determining where to allocate linear VGPRs.