nine: mark the vertex stream when handling DrawIndexedPrimitiveUP
The nine frontend missed setting the vertex stream within the
context->vtxbuf_mask when implementing DrawIndexPrimitiveUP method.
DX9 remarks on this method say that it "supports only a single vertex
stream, which must be declared as stream #0".

The missing state:
(1) causes the update_vertex_elements(_sw) routines to access
    uninitialized elements of vtxbuf_holes_map, causing further such
    accesses to related/dependent data-structures, and
(2) seems to be a probable cause of some crashes (due to accesses to
    uninitialized values) reported at #10557.

The setting of the vertex stream in the vtxbuf_mask must be accompanied
with a recording of the change in the context->changed.vtxbuf. The
present location where the recording occurs is too late; these changes
are meant to be consumed by the function nine_update_state.
Accordingly, move them just before the call to nine_update_state.

These changes mimic a portion of those carried out when handling
SetStreamSource.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/10557
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/11317
Fixes: 9168a376924 ("st/nine: Refactor DrawPrimitiveUp")

Tested-by: Pavel Ondračka <pavel.ondracka@gmail.com>